GCC12 Toolchain Support (with C++20)
GCC12 is the latest version of the GCC toolchain, which importantly adds
support for the C++20 langauge standard, which includes concepts and a
myriad of other useful features. Additonally, it seems to consistently
produce smaller ELF files and .text segments plus increased performance
on the pvrmark benchmark. It also has a fix for a known GCC9 compiler
crash.

- added config.mk.latest.sample as the config file template, using
  GCC version 12.2.0 and Newlib version 4.1.0
- added a patch for the GCC12 compiler
    * required a fix for libobjc configuration
- added a tiny change to _types.h for time_t, as required by Newlib
- tested new toolchain with all KOS examples and verified that nothing
  new is broken
- tested C++20 "concepts" and "constraints" on DC